Updated README and directory structure
